feat: add strict typing compilation

+++ src/app/services/visualize.service.ts
@@ -43,7 +43,7 @@ const GEOCODING_PREVIEW_ID = 'geojson-geocoding-preview';
   providedIn: 'root'
 })
 export class VisualizeService<L, S, M> {
-  private mapInstance: AbstractArlasMapGL;
+  private mapInstance?: AbstractArlasMapGL;
   public fitbounds: Array<Array<number>> = [];
 
   private readonly _isRasterOnMap = signal(false);
@@ -77,7 +77,7 @@ export class VisualizeService<L, S, M> {
     if (urlTemplate.includes('{')) {
       /** Fetch all elements between {} in the template. */
       const regex = new RegExp(/{([^}]+)}/g);
-      const fields = [];
+      const fields = new Array<string>();
       const matches = [...urlTemplate.matchAll(regex)];
       if (matches) {
         matches.filter(m => !!m && Array.isArray(m) && m.length > 1).forEach(m => fields.push(m[1]));
@@ -96,7 +96,7 @@ export class VisualizeService<L, S, M> {
     Observable<string> {
 
     const searchResult = getItem$(elementidentifier, collection, this.collaborativeService);
-    return searchResult.pipe(map(data => flattenedMatchAndReplace(data.hits[0].data, urlTemplate)));
+    return searchResult.pipe(map(data => flattenedMatchAndReplace(data.hits?.[0].data, urlTemplate)));
   }
 
   /** @deprecated Use removeRasters instead. */
@@ -132,6 +132,10 @@ export class VisualizeService<L, S, M> {
    * @param beforeId Insert before a given raster id.
    */
   public addRaster(url: string, maxZoom: number, bounds: Array<number>, id: string, beforeId?: string) {
+    if (!this.mapInstance) {
+      return;
+    }
+
     const layerId = 'raster-source-' + id;
     this.mapFrameworkService.removeLayer(this.mapInstance, layerId);
     this.mapFrameworkService.removeLayer(this.mapInstance, CROSS_LAYER_PREFIX + id);
@@ -167,6 +171,10 @@ export class VisualizeService<L, S, M> {
   }
 
   public addCrossToRemove(lat: number, lng: number, id: string) {
+    if (!this.mapInstance) {
+      return;
+    }
+
     const crossPosition = {
       type: 'FeatureCollection',
       'features': [
@@ -185,10 +193,10 @@ export class VisualizeService<L, S, M> {
       this.notifyRasterRemoved(id);
     });
     this.mapFrameworkService.onLayerEvent('mousemove', this.mapInstance, CROSS_LAYER_PREFIX + id, () => {
-      this.mapFrameworkService.setMapCursor(this.mapInstance, 'pointer');
+      this.mapFrameworkService.setMapCursor(this.mapInstance as AbstractArlasMapGL, 'pointer');
     });
     this.mapFrameworkService.onLayerEvent('mouseleave', this.mapInstance, CROSS_LAYER_PREFIX + id, () => {
-      this.mapFrameworkService.setMapCursor(this.mapInstance, '');
+      this.mapFrameworkService.setMapCursor(this.mapInstance as AbstractArlasMapGL, '');
     });
     this.handlePopup(lat, lng, id);
   }
@@ -198,17 +206,21 @@ export class VisualizeService<L, S, M> {
   }
 
   public handlePopup(lat: number, lng: number, id: string) {
+    if (!this.mapInstance) {
+      return;
+    }
+
     const tooltipMsg = this.translateService.instant('Remove visualisation');
     const popup = this.mapFrameworkService.createPopup(lng, lat, tooltipMsg);
     this.mapFrameworkService.onLayerEvent('mouseenter', this.mapInstance, CROSS_LAYER_PREFIX + id, () => {
-      this.mapFrameworkService.addPopup(this.mapInstance, popup);
+      this.mapFrameworkService.addPopup(this.mapInstance as AbstractArlasMapGL, popup);
     });
     this.mapFrameworkService.onLayerEvent('mouseleave', this.mapInstance, CROSS_LAYER_PREFIX + id, () => {
-      this.mapFrameworkService.removePopup(this.mapInstance, popup);
+      this.mapFrameworkService.removePopup(this.mapInstance as AbstractArlasMapGL, popup);
     });
   }
 
-  public getBoundsAndCenter(idField: string, idValue: string, geometryPath: string, centroidPath: string, collection): Observable<{
+  public getBoundsAndCenter(idField: string, idValue: string, geometryPath: string, centroidPath: string, collection: string): Observable<{
     bounds: Array<Array<number>>;
     center: Array<number>;
     box: BBox;
@@ -226,10 +238,10 @@ export class VisualizeService<L, S, M> {
       .resolveHits([projType.search, search], this.collaborativeService.collaborations, collection, '', filter);
     return searchResult.pipe(
       map(h => {
-        const geomData = getElementFromJsonObject(h.hits[0].data, geometryPath);
-        const centerData = getElementFromJsonObject(h.hits[0].data, centroidPath);
-        const geojsonData = this.getGeojsonFromEsGeom(geomData);
-        const geojsonCenter = this.getGeojsonFromEsGeom(centerData);
+        const geomData = getElementFromJsonObject(h.hits?.[0].data, geometryPath);
+        const centerData = getElementFromJsonObject(h.hits?.[0].data, centroidPath);
+        const geojsonData = this.getGeojsonFromEsGeom(geomData) as GeoJSON.GeoJSON;
+        const geojsonCenter = this.getGeojsonFromEsGeom(centerData) as any;
         const box = bbox(geojsonData);
         const minX = box[0] - 0.1 / 100 * box[0];
         const minY = box[1] - 0.1 / 100 * box[1];
@@ -268,6 +280,10 @@ export class VisualizeService<L, S, M> {
   }
 
   public addGeocodingPreviewLayer(geoJson: any) {
+    if (!this.mapInstance) {
+      return;
+    }
+
     this.mapFrameworkService.removeLayer(this.mapInstance, GEOCODING_PREVIEW_ID);
     const circlePaint: ArlasPaint = {
       'circle-radius': 4,
@@ -284,10 +300,16 @@ export class VisualizeService<L, S, M> {
 
   }
   public removeGeocodingPreviewLayer() {
-    this.mapFrameworkService.removeLayer(this.mapInstance, GEOCODING_PREVIEW_ID);
+    if (this.mapInstance) {
+      this.mapFrameworkService.removeLayer(this.mapInstance, GEOCODING_PREVIEW_ID);
+    }
   }
 
   public handleGeojsonPreview(geojson: any) {
+    if (!this.mapInstance) {
+      return;
+    }
+
     this.addGeocodingPreviewLayer(geojson);
     this.mapFrameworkService.onMapEvent('zoomend', this.mapInstance, () => {
       this.removeGeocodingPreviewLayer();
